split interfaces from repository implementation, rename repository to sqlite specific
